make DefaultGsonReader abstract class containing actual gson object
