+++ src/main/java/org/janelia/saalfeldlab/n5/N5FSReader.java
@@ -38,7 +38,6 @@ import java.nio.file.StandardOpenOption;
 import java.util.HashMap;
 import java.util.stream.Stream;
 
-import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 
@@ -47,7 +46,7 @@ import com.google.gson.JsonElement;
  *
  * @author Stephan Saalfeld
  */
-public class N5FSReader implements DefaultGsonReader {
+public class N5FSReader extends DefaultGsonReader {
 
 	protected static class LockedFileChannel implements Closeable {
 
@@ -96,11 +95,9 @@ public class N5FSReader implements DefaultGsonReader {
 		}
 	}
 
-	protected final String basePath;
-
 	protected static final String jsonFile = "attributes.json";
 
-	protected final Gson gson;
+	protected final String basePath;
 
 	/**
 	 * Opens an {@link N5FSReader} at a given base path with a custom
@@ -111,9 +108,7 @@ public class N5FSReader implements DefaultGsonReader {
 	 */
 	public N5FSReader(final String basePath, final GsonBuilder gsonBuilder) {
 
-		gsonBuilder.registerTypeAdapter(DataType.class, new DataType.JsonAdapter());
-		gsonBuilder.registerTypeAdapter(CompressionType.class, new CompressionType.JsonAdapter());
-		this.gson = gsonBuilder.create();
+		super(gsonBuilder);
 		this.basePath = basePath;
 	}
 
@@ -125,12 +120,6 @@ public class N5FSReader implements DefaultGsonReader {
 	public N5FSReader(final String basePath) {
 
 		this(basePath, new GsonBuilder());
-	}
-
-	@Override
-	public Gson getGson() {
-
-		return gson;
 	}
 
 	@Override
@@ -148,7 +137,7 @@ public class N5FSReader implements DefaultGsonReader {
 			return new HashMap<>();
 
 		try (final LockedFileChannel lockedFileChannel = LockedFileChannel.openForReading(path)) {
-			return GsonAttributesParser.readAttributes(Channels.newReader(lockedFileChannel.getFileChannel(), "UTF-8"), gson);
+			return GsonAttributesParser.readAttributes(Channels.newReader(lockedFileChannel.getFileChannel(), "UTF-8"), getGson());
 		}
 	}
